requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
python-dotenv>=1.0.0
gunicorn>=21.2.0
psycopg2-binary>=2.9.9  # PostgreSQL database adapter
whitenoise>=6.5.0  # Static file serving
django-cors-headers>=4.3.0
